requirements.txt
langchain
python-dotenv
streamlit
openai
googletrans==4.0.0rc1